requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
backoff==2.2.1
botocore==1.40.59
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dotenv==0.9.9
fastapi==0.119.0
frozenlist==1.8.0
gql==4.0.0
graphql-core==3.2.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jmespath==1.0.1
lia-web==0.2.3
multidict==6.7.0
packaging==25.0
pillow==12.0.0
propcache==0.4.1
pydantic==2.12.2
pydantic-settings==2.11.0
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
reportlab==4.4.4
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
strawberry-graphql==0.283.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
yarl==1.22.0
